requirements.txt
flask
tensorflow-cpu
numpy<2.0.0
pillow
gunicorn